WIP(filesystem): first attemp into creating a portable file system #78
